@@ -1,7 +1,10 @@
 from setuptools import setup, find_packages
 
 requirements = [
-    'tensorflow', 'numpy', 'keras', 'keras-bert'
+    'tensorflow',
+    'numpy',
+    'keras',
+    'keras-bert'
 ]
 
 setup(
@@ -13,14 +16,14 @@ setup(
     author='Fleming Kretschmer, Florian Mock, Anton Kriese',
     author_email='fleming.kretschmer@uni-jena.de',
     url='https://github.com/f-kretschmer/bertax',
-    packages=find_packages(),
+    packages=find_packages(include=['bertax']),
     package_data={'bertax': ['resources/big_trainingset_all_fix_classes_selection.h5']},
     entry_points={
         'console_scripts': [
-            'bertax=bertax.bertax:main',
+            'bertax=bertax.bertax:app',
             'bertax-visualize=bertax.visualize:main'
         ]
     },
     install_requires=requirements,
-    python_requires=">=3.7"
+    python_requires=">=3.8"
 )
