@@ -1,166 +1,225 @@
-import numpy as np
-import keras_bert
-from random import randint
 from itertools import product
+from random import randint
+from typing import List, Optional, Tuple, Dict, Iterator, IO
+
+from pathlib import Path
+
+
 import keras
-from logging import info
-import collections
-from typing import List, OrderedDict, Optional, Tuple
+import keras_bert
+import numpy as np
 
-ALPHABET = 'ACGT'
-CLASS_LABELS = OrderedDict([('superkingdom', ['Archaea', 'Bacteria', 'Eukaryota',
-    'Viruses', 'unknown']), ('phylum', ['Actinobacteria', 'Apicomplexa', 'Aquificae',
-    'Arthropoda', 'Artverviricota', 'Ascomycota', 'Bacillariophyta', 'Bacteroidetes',
-    'Basidiomycota', 'Candidatus Thermoplasmatota', 'Chlamydiae', 'Chlorobi',
-    'Chloroflexi', 'Chlorophyta', 'Chordata', 'Crenarchaeota', 'Cyanobacteria',
-    'Deinococcus-Thermus', 'Euglenozoa', 'Euryarchaeota', 'Evosea', 'Firmicutes',
-    'Fusobacteria', 'Gemmatimonadetes', 'Kitrinoviricota', 'Lentisphaerae', 'Mollusca',
-    'Negarnaviricota', 'Nematoda', 'Nitrospirae', 'Peploviricota', 'Pisuviricota',
-    'Planctomycetes', 'Platyhelminthes', 'Proteobacteria', 'Rhodophyta', 'Spirochaetes',
-    'Streptophyta', 'Tenericutes', 'Thaumarchaeota', 'Thermotogae', 'Uroviricota',
-    'Verrucomicrobia', 'unknown']), ('genus', ['Acidilobus', 'Acidithiobacillus',
-    'Actinomyces', 'Actinopolyspora', 'Acyrthosiphon', 'Aeromonas', 'Akkermansia', 'Anas',
-    'Apis', 'Aquila', 'Archaeoglobus', 'Asparagus', 'Aspergillus', 'Astyanax', 'Aythya',
-    'Bdellovibrio', 'Beta', 'Betta', 'Bifidobacterium', 'Botrytis', 'Brachyspira',
-    'Bradymonas', 'Brassica', 'Caenorhabditis', 'Calypte', 'Candidatus Kuenenia',
-    'Candidatus Nitrosocaldus', 'Candidatus Promineofilum', 'Carassius', 'Cercospora',
-    'Chanos', 'Chlamydia', 'Chrysemys', 'Ciona', 'Citrus', 'Clupea', 'Coffea',
-    'Colletotrichum', 'Cottoperca', 'Crassostrea', 'Cryptococcus', 'Cucumis', 'Cucurbita',
-    'Cyanidioschyzon', 'Cynara', 'Cynoglossus', 'Daucus', 'Deinococcus', 'Denticeps',
-    'Desulfovibrio', 'Dictyostelium', 'Drosophila', 'Echeneis', 'Egibacter', 'Egicoccus',
-    'Elaeis', 'Equus', 'Erpetoichthys', 'Esox', 'Euzebya', 'Fervidicoccus', 'Frankia',
-    'Fusarium', 'Gadus', 'Gallus', 'Gemmata', 'Gopherus', 'Gossypium', 'Gouania',
-    'Helianthus', 'Ictalurus', 'Ktedonosporobacter', 'Legionella', 'Leishmania',
-    'Lepisosteus', 'Leptospira', 'Limnochorda', 'Malassezia', 'Manihot', 'Mariprofundus',
-    'Methanobacterium', 'Methanobrevibacter', 'Methanocaldococcus', 'Methanocella',
-    'Methanopyrus', 'Methanosarcina', 'Microcaecilia', 'Modestobacter', 'Monodelphis',
-    'Mus', 'Musa', 'Myripristis', 'Neisseria', 'Nitrosopumilus', 'Nitrososphaera',
-    'Nitrospira', 'Nymphaea', 'Octopus', 'Olea', 'Oncorhynchus', 'Ooceraea',
-    'Ornithorhynchus', 'Oryctolagus', 'Oryzias', 'Ostreococcus', 'Papaver', 'Perca',
-    'Phaeodactylum', 'Phyllostomus', 'Physcomitrium', 'Plasmodium', 'Podarcis', 'Pomacea',
-    'Populus', 'Prosthecochloris', 'Pseudomonas', 'Punica', 'Pyricularia', 'Pyrobaculum',
-    'Quercus', 'Rhinatrema', 'Rhopalosiphum', 'Roseiflexus', 'Rubrobacter', 'Rudivirus',
-    'Salarias', 'Salinisphaera', 'Sarcophilus', 'Schistosoma', 'Scleropages',
-    'Sedimentisphaera', 'Sesamum', 'Solanum', 'Sparus', 'Sphaeramia', 'Spodoptera',
-    'Sporisorium', 'Stanieria', 'Streptomyces', 'Strigops', 'Synechococcus', 'Takifugu',
-    'Thalassiosira', 'Theileria', 'Thermococcus', 'Thermogutta', 'Thermus', 'Tribolium',
-    'Trichoplusia', 'Ustilago', 'Vibrio', 'Vitis', 'Xenopus', 'Xiphophorus',
-    'Zymoseptoria', 'unknown'])])
+from bertax.const import CLASS_LABELS
 
-Record = collections.namedtuple('Record', ['id', 'seq'])
+ALPHABET = "ACGT"
 
-def parse_fasta(fasta) -> List[Record]:
-    records = []
-    id_ = None
+
+def get_file_handle(path: Path) -> IO:
+    if path.suffix == '.gz':
+        import gzip
+        return gzip.open(path, mode='rt')
+    else:
+        return open(path)
+
+
+def parse_fastx(path: Path, handle: IO) -> Iterator[Tuple[str, str]]:
+    if path.suffix == '.gz':
+        suffix = path.suffixes[:-1][-1].lower()
+    else:
+        suffix = path.suffix.lower()
+    if suffix in ['.fasta', '.fna', '.fa']:
+        return parse_fasta(handle)
+    elif suffix in ['.fastq', '.fq']:
+        return parse_fastq(handle)
+    else:
+        raise NotImplementedError(f'No support for parsing files with extension "{suffix}"')
+
+
+def parse_fastq(handle: IO) -> Iterator[Tuple[str, str]]:
+    """Iterate over FASTQ records as string tuples"""
+    header = ''
     seq = ''
-    allowed_characters = set("ACTGYRWSKMDVHBXN")
-    with open(fasta) as f:
-        for line in f:
-            if line.startswith('>'):
-                if (id_ is not None):
-                    records.append(Record(id_, seq))
-                id_ = line[1:].strip()
-                seq = ''
-            else:
-                line_no_whitespace = line.strip()
-                assert set(line_no_whitespace.upper()).issubset(allowed_characters), \
-                    f"{fasta} is not a fasta-file: contains not allowed characters " \
-                    f"{set(line_no_whitespace.upper()).difference(allowed_characters)} in line\n{line_no_whitespace}"
-                seq += line_no_whitespace
-
-        assert id_ is not None, f"{fasta} is not a fasta file: no header detected"
-        records.append(Record(id_, seq))
-    info(f'read in {len(records)} sequences')
-    return records
+    skip = False
+    for line in handle:
+        if skip:
+            skip = False
+            continue
+        line = line.strip()
+        if line == '':
+            continue
+        if line[0] == '@':
+            header = line.replace('@', '').split(maxsplit=1)[0]
+        elif line[0] == '+':
+            yield header, seq
+            skip = True
+        else:
+            seq = line.upper()
 
 
-def seq2kmers(seq, k=3, stride=3, pad=True, to_upper=True):
+def parse_fasta(handle: IO) -> Iterator[Tuple[str, str]]:
+    """Iterate over Fasta records as string tuples.
+    Arguments:
+     - handle - input stream opened in text mode
+    For each record a tuple of two strings is returned, the FASTA title
+    line (without the leading '>' character), and the sequence (with any
+    whitespace removed). The title line is not divided up into an
+    identifier (the first word) and comment or description.
+
+    Source: https://github.com/biopython/biopython/blob/7753efc9fb4a08b845d3f41b3fc790c89743e196/Bio/SeqIO/FastaIO.py#L24
+    """
+    # Skip any text before the first record (e.g. blank lines, comments)
+    for line in handle:
+        if line[0] == ">":
+            title = line[1:].rstrip()
+            break
+    else:
+        # no break encountered - probably an empty file
+        return
+
+    # Main logic
+    # Note, remove trailing whitespace, and any internal spaces
+    # (and any embedded \r which are possible in mangled files
+    # when not opened in universal read lines mode)
+    lines = []
+    for line in handle:
+        if line[0] == ">":
+            yield title, "".join(lines).replace(" ", "").replace("\r", "")
+            lines = []
+            title = line[1:].rstrip()
+            continue
+        lines.append(line.rstrip())
+
+    yield title, "".join(lines).replace(" ", "").replace("\r", "")
+
+
+def seq2kmers(seq: str, k: int = 3, stride: int = 3, pad: bool = True, to_upper: bool = True) -> List[str]:
     """transforms sequence to k-mer sequence.
     If specified, end will be padded so no character is lost"""
-    if (len(seq) < k):
-        return [seq.ljust(k, 'N')] if pad else []
+    if len(seq) < k:
+        return [seq.ljust(k, "N")] if pad else []
     kmers = []
+    i = 0
     for i in range(0, len(seq) - k + 1, stride):
-        kmer = seq[i:i + k]
+        kmer = seq[i: i + k]
         if to_upper:
             kmers.append(kmer.upper())
         else:
             kmers.append(kmer)
     if (pad and len(seq) - (i + k)) % k != 0:
-        kmers.append(seq[i + k:].ljust(k, 'N'))
+        kmers.append(seq[i + k:].ljust(k, "N"))
     return kmers
 
 
-def seq2tokens(seq, token_dict, seq_length=250, max_length=None,
-               k=3, stride=3, window=True, seq_len_like=None):
-    """transforms raw sequence into list of tokens to be used for
-    fine-tuning BERT"""
-    if (max_length is None):
+def seq2tokens(
+        seq: str,
+        token_dict: Dict[str, int],
+        seq_length: int = 250,
+        max_length: int = None,
+        k: int = 3,
+        stride: int = 3,
+        window: bool = True,
+        seq_len_like: int = None,
+) -> Tuple[np.ndarray, np.ndarray]:
+    """Transform raw sequence into list of tokens to be used for fine-tuning BERT
+
+    Args:
+        seq: nucleotide sequence
+        token_dict: Keras BERT base token dict with nucleotide words
+        seq_length:
+        max_length:
+        k:
+        stride:
+        window:
+        seq_len_like:
+
+    Returns:
+        Tuple of: [0] indices int array; [1] zeroed array of length `max_length` for segments
+    """
+    if max_length is None:
         max_length = seq_length
-    if (seq_len_like is not None):
+    if seq_len_like is not None:
         seq_length = min(max_length, np.random.choice(seq_len_like))
-        # open('seq_lens.txt', 'a').write(str(seq_length) + ', ')
-    seq = seq2kmers(seq, k=k, stride=stride, pad=True)
-    if (window):
-        start = randint(0, max(len(seq) - seq_length - 1, 0))
+    kmers = seq2kmers(seq, k=k, stride=stride, pad=True)
+    if window:
+        start = randint(0, max(len(kmers) - seq_length - 1, 0))
         end = start + seq_length - 1
     else:
         start = 0
         end = seq_length
-    indices = [token_dict['[CLS]']] + [token_dict[word]
-                                       if word in token_dict
-                                       else token_dict['[UNK]']
-                                       for word in seq[start:end]]
-    if (len(indices) < max_length):
-        indices += [token_dict['']] * (max_length - len(indices))
+    indices: List[int] = [token_dict["[CLS]"]] + [
+        token_dict[word] if word in token_dict else token_dict["[UNK]"]
+        for word in kmers[start:end]
+    ]
+    if len(indices) < max_length:
+        indices += [token_dict[""]] * (max_length - len(indices))
     else:
         indices = indices[:max_length]
-    segments = [0 for _ in range(max_length)]
-    return [np.array(indices), np.array(segments)]
+    segments = np.zeros(max_length, dtype=int)
+    return np.array(indices), segments
 
-def seq_frames(seq: str, frame_len: int, running_window=False, stride=1):
+
+def seq_frames(seq: str, frame_len: int, running_window=False, stride=1) -> Tuple[List[str], List[Tuple[int, int]]]:
     """returns all windows of seq with a maximum length of `frame_len` and specified stride
     if `running_window` else `frame_len` -- alongside the chunks' positions"""
-    iterator = (range(0, len(seq) - frame_len + 1, stride) if running_window
-                else range(0, len(seq), frame_len))
-    return [seq[i:i + frame_len] for i in iterator], [(i, i + frame_len) for i in iterator]
+    iterator = (
+        range(0, len(seq) - frame_len + 1, stride)
+        if running_window
+        else range(0, len(seq), frame_len)
+    )
+    return [seq[i: i + frame_len] for i in iterator], [
+        (i, i + frame_len) for i in iterator
+    ]
 
-def get_token_dict(alph=ALPHABET, k=3) -> dict:
+
+def get_token_dict(alph: str = ALPHABET, k: int = 3) -> Dict[str, int]:
     """get token dictionary dict generated from `alph` and `k`"""
     token_dict = keras_bert.get_base_dict()
-    for word in [''.join(_) for _ in product(alph, repeat=k)]:
+    for x in product(alph, repeat=k):
+        word = ''.join(x)
         token_dict[word] = len(token_dict)
     return token_dict
+
 
 def process_bert_tokens_batch(batch_x):
     """when `seq2tokens` is used as `custom_encode_sequence`, batches
     are generated as [[input1, input2], [input1, input2], ...]. In
     order to train, they have to be transformed to [input1s,
     input2s] with this function"""
-    return [np.array([_[0] for _ in batch_x]),
-            np.array([_[1] for _ in batch_x])]
+    return np.array([x[0] for x in batch_x]), np.array([x[1] for x in batch_x])
 
-def load_bert(bert_path, compile_=False):
+
+def load_bert(bert_path, compile=False):
     """get bert model from path"""
-    custom_objects = {'GlorotNormal': keras.initializers.glorot_normal,
-                      'GlorotUniform': keras.initializers.glorot_uniform}
+    custom_objects = {
+        "GlorotNormal": keras.initializers.glorot_normal,
+        "GlorotUniform": keras.initializers.glorot_uniform,
+    }
     custom_objects.update(keras_bert.get_custom_objects())
-    model = keras.models.load_model(bert_path, compile=compile_,
-                                    custom_objects=custom_objects)
-    return model
+    return keras.models.load_model(
+        bert_path, compile=compile, custom_objects=custom_objects
+    )
 
-def annotate_predictions(preds: List[np.ndarray],
-                         overwrite_class_labels: Optional[OrderedDict] = None) -> dict:
+
+def annotate_predictions(
+        preds: List[np.ndarray],
+        overwrite_class_labels: Optional[dict] = None
+) -> Dict[str, Dict[str, np.ndarray]]:
     """annotates list of prediction arrays with provided or preset labels"""
-    class_labels = (overwrite_class_labels if overwrite_class_labels is not None
-                    else CLASS_LABELS)
-    return {rank: {l: v.astype(float) for l, v in zip(class_labels[rank], p.transpose())}
-            for rank, p in zip(class_labels, preds)}
+    class_labels = (
+        overwrite_class_labels if overwrite_class_labels is not None else CLASS_LABELS
+    )
+    return {
+        rank: {label: v.astype(float) for label, v in zip(class_labels[rank], p.transpose())}
+        for rank, p in zip(class_labels, preds)
+    }
 
-def best_predictions(preds_annotated: dict) -> List[Tuple[str, float]]:
+
+def best_predictions(
+        preds_annotated: Dict[str, Dict[str, np.ndarray]]
+) -> List[Tuple[str, str, np.ndarray]]:
     """returns best prediction classes alongside predicton confidences"""
     result = []
-    for rank in preds_annotated:
-        best = max(preds_annotated[rank], key=lambda l: preds_annotated[rank][l])
-        result.append((best, preds_annotated[rank][best]))
+    for rank, label_prediction in preds_annotated.items():
+        top_label = max(label_prediction, key=lambda label: label_prediction[label])
+        result.append((rank, top_label, label_prediction[top_label]))
     return result
